requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect2==0.7.0
django==4.1.2
django-check-seo==0.5.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
Pillow==9.2.0
pytz==2022.4
reportlab==3.6.12
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tzdata==2022.5
Unidecode==1.1.2
urllib3==1.26.12
webencodings==0.5.1
